Strict null checks @/contentScript/pipelineProtocol (#8394)
* separate ensureContextMenu

* fix imports

* moves rest of exported backgroudn contextMenu functions to separate files

* fix tests

* overlay

* move remaining files

* move functions to separate files

* fix imports

* lint fixes

* split-up pipelineProtocol

* lint fixes and auto-add

* separate ensureContextMenu

* fix imports

* moves rest of exported backgroudn contextMenu functions to separate files

* fix tests

* overlay

* move remaining files

* move functions to separate files

* fix imports

* lint fixes

* split-up pipelineProtocol

* lint fixes and auto-add
