+++ src/contentScript/messenger/strict/api.ts
@@ -66,3 +66,4 @@ export const insertPanel = getMethod("INSERT_PANEL");
 export const insertButton = getMethod("INSERT_BUTTON");
 export const enableOverlay = getMethod("ENABLE_OVERLAY");
 export const disableOverlay = getMethod("DISABLE_OVERLAY");
+export const runMapArgs = getMethod("RUN_MAP_ARGS");
